model/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
langchain==0.1.20
langchain-community==0.0.38
langchain-chroma==0.1.4
langchain-openai==0.1.7
chromadb==0.5.23
dashscope==1.25.9
openai==1.109.1
pydantic==2.12.5
PyJWT==2.10.1
transformers==4.38.2
sentence-transformers==2.7.0
torch==2.10.0
numpy==1.26.4
pandas==3.0.0
scikit-learn==1.8.0
pdfplumber==0.11.9
python-dotenv==1.2.1
requests==2.32.5
rank_bm25==0.2.2
